Revert "Use double instead of float for extracted features"
This reverts commit 9a50809ebf0f420e72a3c2b4b856dc1a71b9c6b3.

+++ examples/dnn_self_supervised_learning_ex.cpp
@@ -293,10 +293,10 @@ try
     model::feats fnet(layer<5>(net));
     // And we will generate all the features for the training set to train a multiclass SVM
     // classifier.
-    std::vector<matrix<double, 0, 1>> features;
+    std::vector<matrix<float, 0, 1>> features;
     cout << "Extracting features for linear classifier..." << endl;
     features = fnet(training_images, 4 * batch_size);
-    svm_multiclass_linear_trainer<linear_kernel<matrix<double,0,1>>, unsigned long> trainer;
+    svm_multiclass_linear_trainer<linear_kernel<matrix<float,0,1>>, unsigned long> trainer;
     trainer.set_num_threads(std::thread::hardware_concurrency());
     // The most appropriate C setting could be found automatically by using find_max_global().  See the docs for
     // find_max_global() for further information and take particular note of model_selection_ex.cpp.
